requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.2
chromadb>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0